warning suppression

+++ larcv/app/ThreadIO/BatchFillerImage2D.cxx
@@ -42,7 +42,7 @@ namespace larcv {
     _mirrored.clear();
     _mirrored.reserve(batch_size());
 
-    if (!batch_data().dim().empty() && batch_size() != batch_data().dim().front()) {
+    if (!batch_data().dim().empty() && (int)(batch_size()) != batch_data().dim().front()) {
       auto dim = batch_data().dim();
       dim[0] = batch_size();
       this->set_dim(dim);
