+++ larcv/app/ThreadIO/BatchFillerPIDLabel.cxx
@@ -31,7 +31,7 @@ namespace larcv {
 
   void BatchFillerPIDLabel::_batch_begin_()
   {
-    if (!batch_data().dim().empty() && batch_size() != batch_data().dim().front()) {
+    if (!batch_data().dim().empty() && (int)(batch_size()) != batch_data().dim().front()) {
       auto dim = batch_data().dim();
       dim[0] = batch_size();
       this->set_dim(dim);
